Calculate daily cases, daily deaths and daily recoveries from Philippines COVID-19 Data Drop

Calculate daily cases, daily deaths and daily recoveries from Philippines COVID-19 Data Drop

ph_calculate_cases(df)

Arguments

df
A data.frame of cases data pulled from the Philippines COVID-19 Data Drop.

Value

A data.frame with calculated number of cases per day and number of deaths per day and number of recovered per day with structure consistent with CoMo model data structure requirements.

Examples

ph_calculate_cases(df = ph_get_cases())

#> Getting information on Google Drive directory structure for latest available data up to 2020-06-11.

#>        repDate cases deaths recovered
#> 1   2020-01-01     0      0         0
#> 2   2020-01-02     0      0         0
#> 3   2020-01-03     0      0         0
#> 4   2020-01-04     0      0         0
#> 5   2020-01-05     0      0         0
#> 6   2020-01-06     0      0         0
#> 7   2020-01-07     0      0         0
#> 8   2020-01-08     0      0         0
#> 9   2020-01-09     0      0         0
#> 10  2020-01-10     0      0         0
#> 11  2020-01-11     0      0         0
#> 12  2020-01-12     0      0         0
#> 13  2020-01-13     0      0         0
#> 14  2020-01-14     0      0         0
#> 15  2020-01-15     0      0         0
#> 16  2020-01-16     0      0         0
#> 17  2020-01-17     0      0         0
#> 18  2020-01-18     0      0         0
#> 19  2020-01-19     0      0         0
#> 20  2020-01-20     0      0         0
#> 21  2020-01-21     0      0         0
#> 22  2020-01-22     0      0         0
#> 23  2020-01-23     0      0         0
#> 24  2020-01-24     0      0         0
#> 25  2020-01-25     0      0         0
#> 26  2020-01-26     0      0         0
#> 27  2020-01-27     0      0         0
#> 28  2020-01-28     0      0         0
#> 29  2020-01-29     0      0         0
#> 30  2020-01-30     1      0         0
#> 31  2020-01-31     0      0         1
#> 32  2020-02-01     0      1         0
#> 33  2020-02-02     0      0         0
#> 34  2020-02-03     1      1         0
#> 35  2020-02-04     0      0         0
#> 36  2020-02-05     1      0         0
#> 37  2020-02-06     0      0         0
#> 38  2020-02-07     0      0         0
#> 39  2020-02-08     0      0         1
#> 40  2020-02-09     0      0         0
#> 41  2020-02-10     0      0         0
#> 42  2020-02-11     0      0         0
#> 43  2020-02-12     0      0         0
#> 44  2020-02-13     0      0         0
#> 45  2020-02-14     0      0         0
#> 46  2020-02-15     0      0         0
#> 47  2020-02-16     0      0         0
#> 48  2020-02-17     0      0         0
#> 49  2020-02-18     0      0         0
#> 50  2020-02-19     0      0         0
#> 51  2020-02-20     0      0         0
#> 52  2020-02-21     0      0         0
#> 53  2020-02-22     0      0         0
#> 54  2020-02-23     0      0         0
#> 55  2020-02-24     0      0         0
#> 56  2020-02-25     0      0         0
#> 57  2020-02-26     0      0         0
#> 58  2020-02-27     0      0         0
#> 59  2020-02-28     0      0         0
#> 60  2020-02-29     0      0         0
#> 61  2020-03-01     0      0         0
#> 62  2020-03-02     0      0         0
#> 63  2020-03-03     0      0         0
#> 64  2020-03-04     0      0         0
#> 65  2020-03-05     0      0         1
#> 66  2020-03-06     2      0         1
#> 67  2020-03-07     1      0         0
#> 68  2020-03-08     4      0         0
#> 69  2020-03-09    14      0         0
#> 70  2020-03-10     9      0         1
#> 71  2020-03-11    16      3         2
#> 72  2020-03-12     3      3         1
#> 73  2020-03-13    12      4         0
#> 74  2020-03-14    47      5         0
#> 75  2020-03-15    29      6         4
#> 76  2020-03-16     2      5         0
#> 77  2020-03-17    45     12         9
#> 78  2020-03-18    15      9         6
#> 79  2020-03-19    15     14         3
#> 80  2020-03-20    13     16         8
#> 81  2020-03-21    77     14         9
#> 82  2020-03-22    73     22         7
#> 83  2020-03-23    82     30        12
#> 84  2020-03-24    90     23        12
#> 85  2020-03-25    84     23         5
#> 86  2020-03-26    70     33        11
#> 87  2020-03-27   102     38        12
#> 88  2020-03-28   267     32        17
#> 89  2020-03-29   343     29        20
#> 90  2020-03-30   128     38        15
#> 91  2020-03-31   538     25        35
#> 92  2020-04-01   227     26        33
#> 93  2020-04-02   322     24        27
#> 94  2020-04-03   385     17        43
#> 95  2020-04-04    76     26        38
#> 96  2020-04-05   152     21        25
#> 97  2020-04-06   414     18        41
#> 98  2020-04-07   104     15        46
#> 99  2020-04-08   106     18        62
#> 100 2020-04-09   206     18        37
#> 101 2020-04-10   119     18        30
#> 102 2020-04-11   233     21        40
#> 103 2020-04-12   220     14        33
#> 104 2020-04-13   284     16        41
#> 105 2020-04-14   291     11        40
#> 106 2020-04-15   230     18        48
#> 107 2020-04-16   207      2        44
#> 108 2020-04-17   218      8        39
#> 109 2020-04-18   209     13        47
#> 110 2020-04-19   172      8        35
#> 111 2020-04-20   200     13        49
#> 112 2020-04-21   140     10        41
#> 113 2020-04-22   111     11        33
#> 114 2020-04-23   271     14        56
#> 115 2020-04-24   211      7        30
#> 116 2020-04-25   102      6        35
#> 117 2020-04-26   285      6        29
#> 118 2020-04-27   198     16        54
#> 119 2020-04-28   181      3        34
#> 120 2020-04-29   254      8        48
#> 121 2020-04-30   276      7        55
#> 122 2020-05-01   284      4        29
#> 123 2020-05-02   156      5        43
#> 124 2020-05-03   295      8        17
#> 125 2020-05-04   262      7        29
#> 126 2020-05-05   199      5        68
#> 127 2020-05-06   321     12        48
#> 128 2020-05-07   338      5        31
#> 129 2020-05-08   120      2        51
#> 130 2020-05-09   147     14        28
#> 131 2020-05-10   184      4        19
#> 132 2020-05-11   292      6        36
#> 133 2020-05-12   264      5        43
#> 134 2020-05-13   268      8        43
#> 135 2020-05-14   258      9        34
#> 136 2020-05-15   215      8        34
#> 137 2020-05-16   214      7        24
#> 138 2020-05-17   208      2        15
#> 139 2020-05-18   205      3        21
#> 140 2020-05-19   224     11        26
#> 141 2020-05-20   279      2        24
#> 142 2020-05-21   213      5        28
#> 143 2020-05-22   163      6        23
#> 144 2020-05-23   180      5        20
#> 145 2020-05-24   258      1         6
#> 146 2020-05-25   284      2        15
#> 147 2020-05-26   350      5        17
#> 148 2020-05-27   380      1        11
#> 149 2020-05-28   539      3       103
#> 150 2020-05-29  1046      1        79
#> 151 2020-05-30   590      3        57
#> 152 2020-05-31   862      0        85
#> 153 2020-06-01   552      1        78
#> 154 2020-06-02   359      0        71
#> 155 2020-06-03   751      2        33
#> 156 2020-06-04   634      1         3
#> 157 2020-06-05   244      0         4
#> 158 2020-06-06   714      0         0
#> 159 2020-06-07   555      1         2
#> 160 2020-06-08   579      0         0
#> 161 2020-06-09   518      0         1
#> 162 2020-06-10   740      0         0
#> 163 2020-06-11   443      0         0