Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

ph_calculate_rates(df)

Arguments

df
A data.frame of cases data pulled from the Philippines COVID-19 Data Drop.

Value

A data.frame with calculated IHR, IFR, and HFR by age group structure consistent with CoMo model data structure requirements.

Examples

df <- ph_get_cases()

#> Getting information on Google Drive directory structure for latest available data up to 2020-06-10.

ph_calculate_rates(df = df)

#> Getting information on Google Drive directory structure for latest available data up to 2020-06-10.

#>    age_category deaths deathsAdmitted admissions cases         ifr        ihr
#> 1      0-5 y.o.     11              4         71   432 0.025462963 0.16435185
#> 2     5-10 y.o.      2              2         23   257 0.007782101 0.08949416
#> 3    10-15 y.o.      2              2         27   409 0.004889976 0.06601467
#> 4    15-20 y.o.      9              7         68   638 0.014106583 0.10658307
#> 5    20-25 y.o.      3              3        255  1800 0.001666667 0.14166667
#> 6    25-30 y.o.     15              5        454  2853 0.005257624 0.15913074
#> 7    30-35 y.o.     12              7        545  3078 0.003898635 0.17706303
#> 8    35-40 y.o.     26             16        389  2333 0.011144449 0.16673811
#> 9    40-45 y.o.     45             29        378  2150 0.020930233 0.17581395
#> 10   45-50 y.o.     46             29        432  2032 0.022637795 0.21259843
#> 11   50-55 y.o.     74             54        425  1824 0.040570175 0.23300439
#> 12   55-60 y.o.    101             67        466  1677 0.060226595 0.27787716
#> 13   60-65 y.o.    164            113        494  1397 0.117394417 0.35361489
#> 14   65-70 y.o.    147            103        459  1061 0.138548539 0.43261074
#> 15   70-75 y.o.    143             98        353   785 0.182165605 0.44968153
#> 16   75-80 y.o.     97             67        201   444 0.218468468 0.45270270
#> 17   80-85 y.o.     71             52        124   263 0.269961977 0.47148289
#> 18   85-90 y.o.     37             31         65   121 0.305785124 0.53719008
#> 19   90-95 y.o.     15              8         17    37 0.405405405 0.45945946
#> 20  95-100 y.o.      5              5          7    11 0.454545455 0.63636364
#> 21    100+ y.o.      2              1          3   130 0.015384615 0.02307692
#>           hfr
#> 1  0.05633803
#> 2  0.08695652
#> 3  0.07407407
#> 4  0.10294118
#> 5  0.01176471
#> 6  0.01101322
#> 7  0.01284404
#> 8  0.04113111
#> 9  0.07671958
#> 10 0.06712963
#> 11 0.12705882
#> 12 0.14377682
#> 13 0.22874494
#> 14 0.22440087
#> 15 0.27762040
#> 16 0.33333333
#> 17 0.41935484
#> 18 0.47692308
#> 19 0.47058824
#> 20 0.71428571
#> 21 0.33333333