Pull Philippines data on cases information from the publicly available Department of Health COVID-19 Data Drop

A wrapper to googledrive and googlesheets4 functions to pull data from the Philppines' COVID-19 Data Drop resource that is publicly distributed via Google Drive

ph_get_cases(date = Sys.Date())

Arguments

date
Date in <YYYY-MM-DD> format; This is the date up to which extracted data reports to. Default is current date (Sys.Date())

Value

A tibble

Examples

ph_get_cases(date = "2020-04-17")

#> # A tibble: 5,878 x 12
#>    CaseNo   Age AgeGroup Sex   DateRepConf DateRecover DateDied RemovalType
#>    <fct>  <int> <fct>    <fct> <fct>       <fct>       <fct>    <fct>      
#>  1 C1001…    38 35 to 39 Fema… 1/30/2020   "2/8/2020"  ""       "Recovered"
#>  2 C1002…    44 40 to 44 Male  2/3/2020    ""          "2/1/20… "Died"     
#>  3 C1006…    60 60 to 64 Fema… 2/5/2020    "1/31/2020" ""       "Recovered"
#>  4 C1007…    48 45 to 49 Male  3/6/2020    "3/19/2020" ""       "Recovered"
#>  5 C1010…    62 60 to 64 Male  3/6/2020    ""          "3/11/2… "Died"     
#>  6 C1010…    82 80+      Male  4/17/2020   ""          ""       ""         
#>  7 C1010…    57 55 to 59 Fema… 3/7/2020    ""          "3/12/2… "Died"     
#>  8 C1012…    38 35 to 39 Male  3/8/2020    "3/21/2020" ""       "Recovered"
#>  9 C1013…    32 30 to 34 Male  3/8/2020    "4/5/2020"  ""       "Recovered"
#> 10 C1014…    86 80+      Male  3/8/2020    ""          "3/14/2… "Died"     
#> # … with 5,868 more rows, and 4 more variables: DateRepRem <fct>,
#> #   Admitted <fct>, RegionRes <fct>, ProvCityRes <fct>