Pull Philippines data on daily facilities status from the publicly available Department of Health COVID-19 Data Drop

A wrapper to googledrive and googlesheets4 functions to pull data from the Philppines' COVID-19 Data Drop resource that is publicly distributed via Google Drive

ph_get_daily(version = "current", date = NULL)

Arguments

version
A character value specifying whether to get the most currently available dataset ("current") or to get archive data ("archive"). Default to "current"

date
Date in <YYYY-MM-DD> format; This is the date up to which extracted data reports to. Only used when version is set to "archive" otherwise ignored.

Value

A tibble of daily facilities status distributed via the COVID-19 Data Drop

Examples

ph_get_daily()

#> Getting information on Google Drive directory structure of the DoH Data Drop for latest available data up to 2020-06-19.

#> # A tibble: 57,283 x 22
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <int> <chr>  <chr>       <chr>     <chr>      <int> <int>     <int>
#>  1 DOH00000… 53944 PIDDI… 2020-05-20… 2020-05-… 2020-05-2…     0     0         2
#>  2 DOH00000… 55569 PIDDI… 2020-05-21… 2020-05-… 2020-05-2…     0     0         2
#>  3 DOH00000… 57370 PIDDI… 2020-05-22… 2020-05-… 2020-05-2…     0     0         2
#>  4 DOH00000… 59319 PIDDI… 2020-05-23… 2020-05-… 2020-05-2…     0     0         2
#>  5 DOH00000… 61118 PIDDI… 2020-05-24… 2020-05-… 2020-05-2…     0     0         2
#>  6 DOH00000… 63027 PIDDI… 2020-05-25… 2020-05-… 2020-05-2…     0     0         2
#>  7 DOH00000… 65060 PIDDI… 2020-05-26… 2020-05-… 2020-05-2…     0     0         2
#>  8 DOH00000… 67013 PIDDI… 2020-05-27… 2020-05-… 2020-05-2…     0     0         2
#>  9 DOH00000… 68706 PIDDI… 2020-05-28… 2020-05-… 2020-05-2…     0     0         2
#> 10 DOH00000… 70556 PIDDI… 2020-05-29… 2020-05-… 2020-05-2…     0     0         2
#> # … with 57,273 more rows, and 13 more variables: isolbed_o <int>,
#> #   beds_ward_v <int>, beds_ward_o <int>, mechvent_v <int>, mechvent_o <int>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>