Pull Philippines data on testing aggregates from the publicly available Department of Health COVID-19 Data Drop

A wrapper to googledrive and googlesheets4 functions to pull data from the Philppines' COVID-19 Data Drop resource that is publicly distributed via Google Drive

ph_get_tests(date = Sys.Date())

Arguments

date
Date in <YYYY-MM-DD> format; This is the date up to which extracted data reports to. Default is current date (Sys.Date())

Value

A tibble of case information on testing aggregates distributed via the COVID-19 Data Drop

Examples

ph_get_tests(date = "2020-04-17")

#> Warning: cannot open URL 'https://doc-00-18-docs.googleusercontent.com/docs/securesc/ha0ro937gcuc7l7deffksulhg5h7mbp1/flupmkhjdnad4pjg46b8idnt732eunfj/1587277575000/02849385979238948669/*/1rOb0Hznd1Ehr9Le0fvKStQc-p_l4EiNk?e=download': HTTP status was '403 Forbidden'

#> Error in file(file, "rt"): cannot open the connection to 'https://docs.google.com/uc?id=1rOb0Hznd1Ehr9Le0fvKStQc-p_l4EiNk&export=download'