Get Philippines population estimates or projections by 5-year age groups given a year or range of years from the World Population Prospects 2019

Get Philippines population estimates or projections by 5-year age groups given a year or range of years from the World Population Prospects 2019

ph_get_wpp2019(
  file,
  location = "Philippines",
  year = lubridate::year(Sys.Date())
)

Arguments

file
Either a path or a URL to the World Population Prospects 2019 file containing population estimates and projections in 5-year age groups

location
Location to get population estimates/projections for; Default to the "Philippines"

year
A year (numeric) or range of years (YYYY:YYYY) to get population estimates/projections for; Default to current year.

Value

A tibble in tidy format containing population data by 5-year age groups for the entire Philippines and by region from 2015 (estimates) to 2025 (projections)

Examples

link <- "https://population.un.org/wpp/Download/Files/1_Indicators%20(Standard)/CSV_FILES/"
fname <- "WPP2019_PopulationByAgeSex_Medium.csv"
ph_get_wpp2019(file = paste(link, fname, sep = ""))

#>               area year age_categories     total     male   female
#> 984481 Philippines 2020            0-4 10616.342 5450.633 5165.709
#> 984482 Philippines 2020            5-9 11397.952 5846.072 5551.880
#> 984483 Philippines 2020          10-14 10906.801 5578.251 5328.550
#> 984484 Philippines 2020          15-19 10462.894 5407.659 5055.235
#> 984485 Philippines 2020          20-24 10104.334 5191.755 4912.579
#> 984486 Philippines 2020          25-29  9479.780 4807.611 4672.169
#> 984487 Philippines 2020          30-34  8247.197 4166.778 4080.419
#> 984488 Philippines 2020          35-39  7254.730 3644.549 3610.181
#> 984489 Philippines 2020          40-44  6551.963 3297.341 3254.622
#> 984490 Philippines 2020          45-49  5759.264 2883.722 2875.542
#> 984491 Philippines 2020          50-54  5116.997 2538.943 2578.054
#> 984492 Philippines 2020          55-59  4249.923 2075.862 2174.061
#> 984493 Philippines 2020          60-64  3393.186 1614.771 1778.415
#> 984494 Philippines 2020          65-69  2485.685 1137.017 1348.668
#> 984495 Philippines 2020          70-74  1661.429  709.616  951.813
#> 984496 Philippines 2020          75-79   978.190  381.080  597.110
#> 984497 Philippines 2020          80-84   580.536  202.363  378.173
#> 984498 Philippines 2020          85-89   241.033   75.335  165.698
#> 984499 Philippines 2020          90-94    76.064   16.873   59.191
#> 984500 Philippines 2020          95-99    14.988    2.375   12.613
#> 984501 Philippines 2020           100+     1.797    0.219    1.578