Get Philippines population estimates or projections by 5-year age groups given a year or range of years from the World Population Prospects 2019

Get Philippines population estimates or projections by 5-year age groups given a year or range of years from the World Population Prospects 2019

ph_get_wpp2019_pop(
  file,
  location = "Philippines",
  period = lubridate::year(Sys.Date())
)

Arguments

file
Either a path or a URL to the World Population Prospects 2019 file containing population estimates and projections in 5-year age groups

location
Location to get population estimates/projections for; Default to the "Philippines"

period
A year (numeric) or range of years (YYYY:YYYY) to get population estimates/projections for; Default to current year.

Value

A tibble in tidy format containing population data by 5-year age groups for the entire Philippines and by region from 2015 (estimates) to 2025 (projections)

Examples

link <- "https://population.un.org/wpp/Download/Files/1_Indicators%20(Standard)/CSV_FILES/"
fname <- "WPP2019_PopulationByAgeSex_Medium.csv"
ph_get_wpp2019_pop(file = paste(link, fname, sep = ""))

#> # A tibble: 21 x 6
#>    area         year age_category    total     male  female
#>    <chr>       <int> <chr>           <dbl>    <dbl>   <dbl>
#>  1 Philippines  2020 0-4          10616342 5450633  5165709
#>  2 Philippines  2020 5-9          11397952 5846072  5551880
#>  3 Philippines  2020 10-14        10906801 5578251  5328550
#>  4 Philippines  2020 15-19        10462894 5407659  5055235
#>  5 Philippines  2020 20-24        10104334 5191755  4912579
#>  6 Philippines  2020 25-29         9479780 4807611  4672169
#>  7 Philippines  2020 30-34         8247197 4166778. 4080419
#>  8 Philippines  2020 35-39         7254730 3644549  3610181
#>  9 Philippines  2020 40-44         6551963 3297341  3254622
#> 10 Philippines  2020 45-49         5759264 2883722  2875542
#> # … with 11 more rows