Calculate daily cases and daily deaths from Philippines COVID-19 Data Drop

Calculate daily cases and daily deaths from Philippines COVID-19 Data Drop

ph_calculate_cases(date = Sys.Date())

Arguments

date
Date in <YYYY-MM-DD> format; This is the date up to which extracted data reports to. Default is current date (Sys.Date())

Value

A data.frame with calculated number of cases per and number of deaths per day with structure consistent with CoMo model data structure requirements.

Examples

ph_calculate_cases(date = "2020-04-17")

#>        repDate cases deaths
#> 1   2020-01-01     0      0
#> 2   2020-01-02     0      0
#> 3   2020-01-03     0      0
#> 4   2020-01-04     0      0
#> 5   2020-01-05     0      0
#> 6   2020-01-06     0      0
#> 7   2020-01-07     0      0
#> 8   2020-01-08     0      0
#> 9   2020-01-09     0      0
#> 10  2020-01-10     0      0
#> 11  2020-01-11     0      0
#> 12  2020-01-12     0      0
#> 13  2020-01-13     0      0
#> 14  2020-01-14     0      0
#> 15  2020-01-15     0      0
#> 16  2020-01-16     0      0
#> 17  2020-01-17     0      0
#> 18  2020-01-18     0      0
#> 19  2020-01-19     0      0
#> 20  2020-01-20     0      0
#> 21  2020-01-21     0      0
#> 22  2020-01-22     0      0
#> 23  2020-01-23     0      0
#> 24  2020-01-24     0      0
#> 25  2020-01-25     0      0
#> 26  2020-01-26     0      0
#> 27  2020-01-27     0      0
#> 28  2020-01-28     0      0
#> 29  2020-01-29     0      0
#> 30  2020-01-30     1      0
#> 31  2020-01-31     0      0
#> 32  2020-02-01     0      1
#> 33  2020-02-02     0      0
#> 34  2020-02-03     1      0
#> 35  2020-02-04     0      0
#> 36  2020-02-05     1      0
#> 37  2020-02-06     0      0
#> 38  2020-02-07     0      0
#> 39  2020-02-08     0      0
#> 40  2020-02-09     0      0
#> 41  2020-02-10     0      0
#> 42  2020-02-11     0      0
#> 43  2020-02-12     0      0
#> 44  2020-02-13     0      0
#> 45  2020-02-14     0      0
#> 46  2020-02-15     0      0
#> 47  2020-02-16     0      0
#> 48  2020-02-17     0      0
#> 49  2020-02-18     0      0
#> 50  2020-02-19     0      0
#> 51  2020-02-20     0      0
#> 52  2020-02-21     0      0
#> 53  2020-02-22     0      0
#> 54  2020-02-23     0      0
#> 55  2020-02-24     0      0
#> 56  2020-02-25     0      0
#> 57  2020-02-26     0      0
#> 58  2020-02-27     0      0
#> 59  2020-02-28     0      0
#> 60  2020-02-29     0      0
#> 61  2020-03-01     0      0
#> 62  2020-03-02     0      0
#> 63  2020-03-03     0      0
#> 64  2020-03-04     0      0
#> 65  2020-03-05     0      0
#> 66  2020-03-06     2      0
#> 67  2020-03-07     1      0
#> 68  2020-03-08     4      0
#> 69  2020-03-09    14      0
#> 70  2020-03-10     9      0
#> 71  2020-03-11    16      3
#> 72  2020-03-12     3      3
#> 73  2020-03-13    12      4
#> 74  2020-03-14    47      4
#> 75  2020-03-15    29      5
#> 76  2020-03-16     2      3
#> 77  2020-03-17    45     11
#> 78  2020-03-18    15      8
#> 79  2020-03-19    15     12
#> 80  2020-03-20    13     12
#> 81  2020-03-21    77     10
#> 82  2020-03-22    73     14
#> 83  2020-03-23    82     18
#> 84  2020-03-24    90     12
#> 85  2020-03-25    84     12
#> 86  2020-03-26    70     18
#> 87  2020-03-27   102     21
#> 88  2020-03-28   267     15
#> 89  2020-03-29   343     17
#> 90  2020-03-30   128     10
#> 91  2020-03-31   538     13
#> 92  2020-04-01   227     12
#> 93  2020-04-02   322     12
#> 94  2020-04-03   385     14
#> 95  2020-04-04    76     12
#> 96  2020-04-05   152      6
#> 97  2020-04-06   414     10
#> 98  2020-04-07   104      3
#> 99  2020-04-08   106      8
#> 100 2020-04-09   206      4
#> 101 2020-04-10   119      5
#> 102 2020-04-11   233      3
#> 103 2020-04-12   220      4
#> 104 2020-04-13   284      2
#> 105 2020-04-14   291      1
#> 106 2020-04-15   230      1
#> 107 2020-04-16   207      0
#> 108 2020-04-17   218      0