Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

ph_calculate_rates(df)

Arguments

df
A data.frame of cases data pulled from the Philippines COVID-19 Data Drop.

Value

A data.frame with calculated IHR, IFR, and HFR by age group structure consistent with CoMo model data structure requirements.

Examples

df <- ph_get_cases()

#> Getting information on Google Drive directory structure for latest available data up to 2020-06-11.

ph_calculate_rates(df = df)

#> Getting information on Google Drive directory structure for latest available data up to 2020-06-11.

#>    age_category deaths deathsAdmitted admissions cases         ifr        ihr
#> 1      0-5 y.o.     11              4         70   437 0.025171625 0.16018307
#> 2     5-10 y.o.      2              2         23   262 0.007633588 0.08778626
#> 3    10-15 y.o.      2              2         27   412 0.004854369 0.06553398
#> 4    15-20 y.o.      9              7         70   652 0.013803681 0.10736196
#> 5    20-25 y.o.      3              3        261  1844 0.001626898 0.14154013
#> 6    25-30 y.o.     15              5        458  2935 0.005110733 0.15604770
#> 7    30-35 y.o.     12              7        561  3135 0.003827751 0.17894737
#> 8    35-40 y.o.     27             16        394  2379 0.011349306 0.16561580
#> 9    40-45 y.o.     45             29        382  2185 0.020594966 0.17482838
#> 10   45-50 y.o.     47             30        436  2061 0.022804464 0.21154779
#> 11   50-55 y.o.     76             56        431  1863 0.040794418 0.23134729
#> 12   55-60 y.o.    102             67        471  1703 0.059894304 0.27657076
#> 13   60-65 y.o.    164            113        497  1424 0.115168539 0.34901685
#> 14   65-70 y.o.    148            104        465  1074 0.137802607 0.43296089
#> 15   70-75 y.o.    144            100        356   792 0.181818182 0.44949495
#> 16   75-80 y.o.     99             67        206   453 0.218543046 0.45474614
#> 17   80-85 y.o.     71             52        124   266 0.266917293 0.46616541
#> 18   85-90 y.o.     37             31         65   123 0.300813008 0.52845528
#> 19   90-95 y.o.     15              8         17    37 0.405405405 0.45945946
#> 20  95-100 y.o.      5              5          7    11 0.454545455 0.63636364
#> 21    100+ y.o.      2              1          3   127 0.015748031 0.02362205
#>           hfr
#> 1  0.05714286
#> 2  0.08695652
#> 3  0.07407407
#> 4  0.10000000
#> 5  0.01149425
#> 6  0.01091703
#> 7  0.01247772
#> 8  0.04060914
#> 9  0.07591623
#> 10 0.06880734
#> 11 0.12993039
#> 12 0.14225053
#> 13 0.22736419
#> 14 0.22365591
#> 15 0.28089888
#> 16 0.32524272
#> 17 0.41935484
#> 18 0.47692308
#> 19 0.47058824
#> 20 0.71428571
#> 21 0.33333333