Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

Calculate infection fatality rate (IFR), infection hospitalisation rate, and hospitalisation fatality rate (HFR) from Philippines data

ph_calculate_rates(df)

Arguments

df
A data.frame of cases data pulled from the Philippines COVID-19 Data Drop.

Value

A data.frame with calculated IHR, IFR, and HFR by age group structure consistent with CoMo model data structure requirements.

Examples

df <- ph_get_cases()

#> Getting information on Google Drive directory structure of the DoH Data Drop for latest available data up to 2020-06-23.

ph_calculate_rates(df = df)

#> # A tibble: 21 x 8
#>    age_category deaths deathsAdmitted admissions cases     ifr    ihr     hfr
#>    <fct>         <dbl>          <dbl>      <dbl> <dbl>   <dbl>  <dbl>   <dbl>
#>  1 0-5 y.o.         14              7         93   579 0.0242  0.161  0.0753 
#>  2 5-10 y.o.         3              2         32   372 0.00806 0.0860 0.0625 
#>  3 10-15 y.o.        2              2         41   558 0.00358 0.0735 0.0488 
#>  4 15-20 y.o.        9              9        104   895 0.0101  0.116  0.0865 
#>  5 20-25 y.o.        4              3        355  2640 0.00152 0.134  0.00845
#>  6 25-30 y.o.       21              7        591  4097 0.00513 0.144  0.0118 
#>  7 30-35 y.o.       13              7        689  4175 0.00311 0.165  0.0102 
#>  8 35-40 y.o.       34             21        503  3127 0.0109  0.161  0.0417 
#>  9 40-45 y.o.       50             35        453  2804 0.0178  0.162  0.0773 
#> 10 45-50 y.o.       55             34        521  2674 0.0206  0.195  0.0653 
#> # … with 11 more rows